apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
python-multipart==0.0.18
httpx==0.28.0
litellm==1.54.0
redis==5.2.0
langgraph==0.2.58
python-dateutil==2.9.0
